Control.Monad.Coproduct

Description

Ideal Monad Coproduct

newtype ((m0 :: Type -> Type) :+ (n0 :: Type -> Type)) a

Coproduct of (impure parts of) two Monads.

As the coproduct of Isolated

Given Isolated m0 and Isolated n0, the functor m0 :+ n0 is Isolated too. In other words, given two Monads Unite m0 and Unite n0, this type constructs a new Monad (Unite (m0 :+ n0)).

Furthermore, as the name suggests, Unite (m0 :+ n0) is the coproduct of Unite m0 and Unite n0 as a Monad.

hoistUnite inject1 :: Unite m0 ~> Unite (m0 :+ n0) is a monad morphism

hoistUnite inject2 :: Unite n0 ~> Unite (m0 :+ n0) is a monad morphism

eitherMonad mt nt :: (m0 :+ n0) ~> t is an impure monad morphism, given (mt :: m0 ~> t) and (nt :: n0 ~> t) are impure monad morphisms.

Any impure monad morphisms (m0 :+ n0) ~> t can be uniquely rewritten as (eitherMonad mt nt).

Here, a natural transformation nat :: f ~> g is an impure monad morphism means f is an Isolated, g is a Monad, and nat becomes a monad morphism when combined with pure as below.

either pure nat . runUnite :: Unite f ~> g

As the coproduct of MonadIdeal

Given MonadIdeal m0 and MonadIdeal n0, the functor m0 :+ n0 is MonadIdeal too. It is the coproduct of m0 and n0 as a MonadIdeal.

inject1 :: m0 ~> (m0 :+ n0) is a MonadIdeal morphism

inject2 :: n0 ~> (m0 :+ n0) is a MonadIdeal morphism

(mt |||| nt) :: (m0 :+ n0) ~> t0 is a MonadIdeal morphism, given mt, nt are MonadIdeal morphisms.

Any MonadIdeal morphism (m0 :+ n0) ~> t0 can be uniquely rewritten as (mt |||| nt).

Here, a MonadIdeal morphism is a natural transformation nat between MonadIdeal such that preserves idealBind.

nat (idealBind ma k) = idealBind (nat ma) (hoistIdeal nat . k)

Constructors

Coproduct

Fields

runCoproduct :: Either (Mutual Either m0 n0 a) (Mutual Either n0 m0 a)

inject1 :: forall m0 a (n0 :: Type -> Type). Functor m0 => m0 a -> (m0 :+ n0) a

inject2 :: forall n0 a (m0 :: Type -> Type). Functor n0 => n0 a -> (m0 :+ n0) a

(||||) :: MonadIdeal t => (forall a. m0 a -> t a) -> (forall a. n0 a -> t a) -> (m0 :+ n0) b -> t b

eitherMonad :: (Isolated m0, Isolated n0, Monad t) => (forall a. m0 a -> t a) -> (forall a. n0 a -> t a) -> (m0 :+ n0) b -> t b

Mutual recursion for ideal monad coproducts

newtype Mutual (p :: Type -> Type -> Type) (m :: Type -> Type) (n :: Type -> Type) a

Constructors

Mutual

Fields

runMutual :: m (p a (Mutual p n m a))