Control.Monad.Ideal

Description

Ideal Monads

class (Bind m, Isolated m) => MonadIdeal (m :: Type -> Type) where

m is the "ideal part" of an ideal monad.

Laws

Methods inherited from superclasses must be equivalent to the canocical ones.

(>>-) === bindDefault

impureBind === impureBindDefault

Methods

idealBind :: m a -> (a -> Ideal m b) -> m b infixl 1

idealize :: MonadIdeal m => m (Ideal m a) -> m a

type Ideal = Unite

Ideal monad is a special case of Unite m

ideal :: Either a (m a) -> Ideal m a

Constructor of ideal = Unite for backward compatibility

runIdeal :: Ideal m a -> Either a (m a)

Deconstructor of ideal = Unite for backward compatibility

hoistIdeal :: (forall a. m a -> n a) -> Ideal m b -> Ideal n b

Alias of hoistUnite for naming consistently

destroyIdeal :: (m a -> a) -> Ideal m a -> a

bindDefault :: MonadIdeal m => m a -> (a -> m b) -> m b

MonadIdeal implies Bind

impureBindDefault :: MonadIdeal m => m a -> (a -> Unite m b) -> Unite m b

MonadIdeal implies Isolated

Relation between MonadIdeal, Bind, and Isolated

MonadIdeal is a requirement stronger than both of Bind and Isolated. In fact, these subset relations hold.

Every MonadIdeal is Bind

Every MonadIdeal is Isolated

These are strict subset relation and neither of three classes can be dropped.

NotOne is both Bind and Isolated, but not MonadIdeal.

NullBind c is Bind but can't be Isolated, because Unite (NullBind c) can't be a Monad in a compatible way.

newtype NullBind c a = NullBind (Maybe c)
instance Bind (NullBind c a) where
  _ >>- _ = NullBind Nothing

Toggle shown below is Isolated, but can't be a Bind in a compatible way.

newtype Toggle a = Toggle a
  deriving Functor

instance Isolated Toggle where
  impureBind (Toggle a) k = case k a of
    Unite (Left b)           -> Unite (Right (Toggle b))
    Unite (Right (Toggle b)) -> Unite (Left b)