Control.Monad.Isolated

Impure part isolated from a Monad

class Functor m => Isolated (m :: Type -> Type) where

Isolated m is a Functor which can be thought of as an impure part of a Monad.

Examples

Proxy is Isolated by being same to the Nothing part of the Maybe monad.

NotOne is Isolated by being the list monad ('[]') minus singleton lists, the pure part of the list monad.

Non-example

Not every Monad can be isolated its pure and impure parts as the sum of functors. For example, the reader monad cannot be written as a sum of two functors.

Laws

impureBind must be implemented so that the Monad (Unite m) instance derived from it is lawful.

return a = Unite (Left a)

Unite (Left a) >>= k = k a
Unite (Right ma) >>= k = ma `impureBind` k

Translating the Monad laws on Unite m in terms of impureBind, the following equations are the Isolated laws on its own.

(Right identity)

ma `impureBind` Unite . Left === Unite (Right ma)

(Associativity)

(ma `impureBind` f) `impureBind` g === ma impureBind a -> either g (`impureBind` g) (runUnite fa)

Methods

impureBind :: m a -> (a -> Unite m b) -> Unite m b infixl 1

(Re)Unite a Monad from pure and impure parts

newtype Unite (f :: Type -> Type) a

Constructors

Unite

Fields

runUnite :: Either a (f a)

hoistUnite :: (forall a. f a -> g a) -> Unite f b -> Unite g b