Data.Functor.KeepLeft

Description

Documentation

newtype KeepLeft c a

Another choices of instances for Const. KeepLeft c have Apply instance isomorphic to Const (First c), which can also (exceptionally) have Bind instance.

The most used constant functor Const lacks the instance of Bind, due to incompatibility between Bind and Apply.

instance (Semigroup c) => Apply (Const c)

While any Semigroup c instance yields lawful Apply (Const c) instance, large number of them do not have Bind instance compatible to Apply. One of few exceptional Semigroup instance is ones isomorphic to First c semigroup, whose semigroup operation is "always use the left operand of <>."

(<>) :: c -> c -> c
x <> _ = x

Constructors

KeepLeft

Fields

getKeepLeft :: c